+++ docs/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Release History
 
+## 3.0.5 (2020-05-22)
+
+**Added**
+
+- feat: each testcase has an unique id in uuid4 format
+- feat: add default header `HRUN-Request-ID` for each testcase #721
+- feat: builtin allure report
+- feat: dump log for each testcase
+
+**Fixed**
+
+- fix: ensure referenced testcase share the same session
+
+**Changed**
+
+- change: remove default added `-s` option for hrun
+
 ## 3.0.4 (2020-05-19)
 
 **Added**